+++ Reciplease/Recipe.swift
@@ -13,9 +13,9 @@ class Recipe {
     let imageUrl: String
     let recipeUrl: String
     let ingredients: [Ingredient]
-    let mark: Float
+    let mark: Int
     
-    init(name: String, image: String, recipe: String, ingredients: [Ingredient], mark: Float) {
+    init(name: String, image: String, recipe: String, ingredients: [Ingredient], mark: Int) {
         self.name = name
         imageUrl = image
         recipeUrl = recipe
